+++ src/meltano/api/controllers/orchestrations.py
@@ -1,4 +1,7 @@
+import asyncio
+import json
 import logging
+import psutil
 
 from flask import Blueprint, request, url_for, jsonify, make_response, Response
 
@@ -10,6 +13,7 @@ from meltano.core.plugin.settings_service import (
     PluginSettingValueSource,
 )
 from meltano.core.plugin_discovery_service import PluginDiscoveryService
+from meltano.core.plugin_invoker import invoker_factory
 from meltano.core.plugin_install_service import PluginInstallService
 from meltano.core.project import Project
 from meltano.core.project_add_service import ProjectAddService
@@ -124,7 +128,7 @@ def run():
 @orchestrationsBP.route("/<plugin_ref:plugin_ref>/configuration", methods=["GET"])
 def get_plugin_configuration(plugin_ref) -> Response:
     """
-    endpoint for getting a plugin's configuration
+    Endpoint for getting a plugin's configuration
     """
     project = Project.find()
     settings = PluginSettingsService(project)
@@ -144,7 +148,7 @@ def get_plugin_configuration(plugin_ref) -> Response:
 @orchestrationsBP.route("/<plugin_ref:plugin_ref>/configuration", methods=["PUT"])
 def save_plugin_configuration(plugin_ref) -> Response:
     """
-    endpoint for persisting a plugin configuration
+    Endpoint for persisting a plugin configuration
     """
     project = Project.find()
     payload = request.get_json()
@@ -164,10 +168,57 @@ def save_plugin_configuration(plugin_ref) -> Response:
     return jsonify(settings.as_config(db.session, plugin_ref, redacted=True))
 
 
+@orchestrationsBP.route("/<plugin_ref:plugin_ref>/configuration/test", methods=["POST"])
+def test_plugin_configuration(plugin_ref) -> Response:
+    """
+    Endpoint for testing a plugin configuration's valid connection
+    """
+    project = Project.find()
+    payload = request.get_json()
+    config_service = ConfigService(project)
+    plugin = config_service.find_plugin(plugin_ref.name, plugin_ref.type)
+
+    async def test_stream(tap_stream) -> bool:
+        while not tap_stream.at_eof():
+            message = await tap_stream.readline()
+            json_dict = json.loads(message)
+            if json_dict["type"] == "RECORD":
+                return True
+
+        return False
+
+    async def test_extractor(config={}):
+        try:
+            invoker = invoker_factory(project, plugin, prepare_with_session=db.session)
+            # overlay the config on top of the loaded configuration
+            invoker.plugin_config = {
+                **invoker.plugin_config,
+                **PluginSettingsService.unredact(config),  # remove all redacted values
+            }
+
+            invoker.prepare(db.session)
+            process = await invoker.invoke_async(stdout=asyncio.subprocess.PIPE)
+            return await test_stream(process.stdout)
+        except Exception as err:
+            # if anything happens, this is not successful
+            return False
+        finally:
+            try:
+                if process:
+                    psutil.Process(process.pid).terminate()
+            except Exception as err:
+                logging.debug(err)
+
+    loop = asyncio.get_event_loop()
+    success = loop.run_until_complete(test_extractor(payload))
+
+    return jsonify({"is_success": success}), 200
+
+
 @orchestrationsBP.route("/pipeline_schedules", methods=["GET"])
 def get_pipeline_schedules():
     """
-    endpoint for getting the pipeline schedules
+    Endpoint for getting the pipeline schedules
     """
     project = Project.find()
     schedule_service = ScheduleService(project)
@@ -185,7 +236,7 @@ def get_pipeline_schedules():
 @orchestrationsBP.route("/pipeline_schedules", methods=["POST"])
 def save_pipeline_schedule() -> Response:
     """
-    endpoint for persisting a pipeline schedule
+    Endpoint for persisting a pipeline schedule
     """
     incoming = request.get_json()
     # Airflow requires alphanumeric characters, dashes, dots and underscores exclusively